+++ src/main/java/io/github/wysohn/realeconomy/manager/banking/account/TradingAccount.java
@@ -23,7 +23,7 @@ public class TradingAccount implements IAccount, IAssetHolder {
     }
 
     @Override
-    public Collection<Asset> removeAsset(AssetSignature signature, int amount) {
+    public Collection<Asset> removeAsset(AssetSignature signature, double amount) {
         return AssetUtil.removeAsset(ownedAssets, signature, amount);
     }
